+++ packages/firebase-utils/src/storage/index.ts
@@ -1,2 +1,5 @@
 export * from './bucket'
 export * from './gcs-object'
+
+export * from './upload-file'
+export * from './upload-json'