+++ packages/ts-core/src/crypto/index.ts
@@ -1,3 +1,4 @@
 export * from './uuid'
 export * from './md5'
 export * from './object-hash'
+export * from './sha256'