+++ packages/ts-core/src/data/object/index.ts
@@ -1,3 +1,4 @@
 export * from './remove-fields'
 export * from './map-fields'
 export * from './byte-size'
+export * from './ndjson'