+++ packages/ts-core/src/data/object/ndjson.ts
@@ -1,0 +1,5 @@
+
+
+export function toNDJSON(objArray: unknown[]): Buffer {
+	return Buffer.from(objArray.map(obj => JSON.stringify(obj)).join('\n'), 'utf-8')
+}